test: Fix conversion warnings in GTest

+++ test/unittests/test_intx.cpp
@@ -245,8 +245,8 @@ TEST(uint256, exp)
 TEST(uint256, count_significant_bytes)
 {
     auto w = count_significant_words<uint8_t>(1_u256 << 113);
-    EXPECT_EQ(w, 15);
-    EXPECT_EQ(count_significant_words<uint8_t>(0_u256), 0);
+    EXPECT_EQ(w, 15u);
+    EXPECT_EQ(count_significant_words<uint8_t>(0_u256), 0u);
 }
 
 template <typename T>
@@ -408,10 +408,10 @@ TYPED_TEST(uint_test, count_significant_words_32)
     constexpr auto csw = count_significant_words<uint32_t, TypeParam>;
 
     TypeParam x;
-    EXPECT_EQ(csw(x), 0);
+    EXPECT_EQ(csw(x), 0u);
 
     x = 1;
-    for (unsigned s = 0; s < sizeof(TypeParam) * 8; ++s)
+    for (size_t s = 0; s < sizeof(TypeParam) * 8; ++s)
         EXPECT_EQ(csw(x << s), s / 32 + 1);
 }
 
@@ -420,10 +420,10 @@ TYPED_TEST(uint_test, count_significant_words_64)
     constexpr auto csw = count_significant_words<uint64_t, TypeParam>;
 
     TypeParam x;
-    EXPECT_EQ(csw(x), 0);
+    EXPECT_EQ(csw(x), 0u);
 
     x = 1;
-    for (unsigned s = 0; s < sizeof(TypeParam) * 8; ++s)
+    for (size_t s = 0; s < sizeof(TypeParam) * 8; ++s)
         EXPECT_EQ(csw(x << s), s / 64 + 1);
 }
